use cobra for manage cmd
