+++ README.md
@@ -4,11 +4,10 @@
 # run (dev)
 go mod tidy
 
-Node 1 (first nodes):
-go run cmd/main.go -rpcPort="8080" -listenAddress="localhost:8000" -peerAddresses="localhost:8001"
+go run cmd/main.go --node <node name> 
 
-Node 2:
-go run cmd/main.go -rpcPort="8081" -listenAddress="localhost:8001" -peerAddresses="localhost:8000"
+exaple:
+go run cmd/main.go --node node3
 
 # Build
 go build -o bin/mmn ./cmd
@@ -16,3 +15,5 @@ go build -o bin/mmn ./cmd
 # uses cases
 Mezon -> (auto gen wallet) => user has a wallet
 Mezon (wallet) -> create and sign transaction -> send rpc -> mmn node verify commit and broadcast to nodes.
+
+
