+++ blockstore/blocks/0.json
@@ -1,0 +1,1 @@
+{"Slot":0,"PrevHash":[126,156,143,104,93,45,111,227,38,227,80,121,128,213,173,28,62,34,134,131,66,40,123,8,223,9,166,194,9,195,169,111],"Entries":[{"num_hashes":5,"hash":[131,147,202,72,29,165,238,146,92,144,155,32,121,46,131,28,197,243,232,219,233,144,194,208,17,211,113,56,162,178,223,34],"transactions":null,"tick":true},{"num_hashes":5,"hash":[24,204,122,200,204,209,1,170,118,63,126,148,19,140,161,20,144,91,212,93,115,130,228,109,181,1,3,52,51,96,12,148],"transactions":null,"tick":true},{"num_hashes":5,"hash":[223,28,181,230,102,111,90,183,160,30,160,31,98,83,168,233,231,134,149,170,108,38,253,155,118,209,29,232,78,28,126,147],"transactions":null,"tick":true}],"LeaderID":"6a4dd9b6efe0fc8f125be331735b0e33239e24f02c84e555ade9ea50bd1369db","Timestamp":"2025-08-11T17:22:45.4234438+07:00","Hash":[115,145,141,38,151,167,80,253,242,250,50,126,160,32,80,191,242,182,71,194,63,56,222,111,250,0,142,147,18,47,5,97],"Signature":"+n3vKzgG4rPdlWbbYi2xdo2DUg1eVKb/NFeExf2XigXeKhbAIs4yHesWc5V9iOPQYTNFQw9vbFQnx1YHzbMKCg==","Status":0}
