+++ main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"mmn/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
